Fix kblayout file names
