+++ toshy_common/kblayout_context.py
@@ -1,133 +1,176 @@
 #!/usr/bin/env python3
+"""
+Keyboard-layout context coordinator for Toshy.
+
+    toshy_common/kblayout_context.py
+
+The thin runtime layer that ties the three building blocks together: it watches
+the active layout (kblayout_detect), analyzes each new layout against a
+reference (kblayout_analyze), and hands the resulting keycode->keycode
+correction map to the keymapper. This is the module the config/runtime wires
+up. The detector and analyzer never know about each other; this coordinator is
+the only thing that knows both, and it stays deliberately small.
+
+The keymapper hand-off is an injected callback (apply_correction_map) rather
+than a direct keymapper call, so this module does not depend on the keymapper's
+correction-map setter and can be built and tested before that setter exists.
+The callback runs on the detector's watcher thread — the same threading
+contract kblayout_detect documents — so a consumer that needs the main thread
+marshals there itself.
+"""
+
+__version__ = '20260604'
 
 import os
 import sys
 import signal
-import platform
-from xkbcommon import xkb
+import threading
 
-# Set up logging
-VERBOSE = True
-FLUSH = True
 
-def debug(*args, ctx="DD"):
-    if not VERBOSE:
-        return
-    if len(args) == 0 or (len(args) == 1 and args[0] == ""):
-        print("", flush=FLUSH)
-        return
-    print(f"({ctx})", *args, flush=FLUSH)
+_toshy_root = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
+if _toshy_root not in sys.path:
+    sys.path.insert(0, _toshy_root)
 
-def warn(*args, ctx="WW"):
-    print(f"({ctx})", *args, flush=FLUSH)
+from toshy_common.kblayout_common import format_layout
+from toshy_common.kblayout_detect import KeyboardLayoutDetector
+from toshy_common.kblayout_analyze import KeyboardLayoutAnalyzer
 
-def error(*args, ctx="EE"):
-    print(f"({ctx})", *args, flush=FLUSH)
 
-def info(*args, ctx="--"):
-    print(f"({ctx})", *args, flush=FLUSH)
+def _warn(message):
+    """Emit a warning to stderr (kept simple and dependency-free)."""
+    print(f'(WW) kblayout_context: {message}', file=sys.stderr)
 
-# Signal handling
-def signal_handler(sig, frame):
-    if sig in (signal.SIGINT, signal.SIGQUIT):
-        print('\n')
-        debug(f'Signal {sig} received. Cleaning up and exiting.\n')
-        sys.exit(0)
 
-# Set up signal handling
-if platform.system() != 'Windows':
-    signal.signal(signal.SIGINT, signal_handler)
-    signal.signal(signal.SIGQUIT, signal_handler) 
-    signal.signal(signal.SIGHUP, signal_handler)
-else:
-    error(f'This is only meant to run on Linux. Exiting.')
-    sys.exit(1)
+class KeyboardLayoutContext:
+    """Maintains the active layout's correction map and keeps it in sync.
 
-class KeymapAnalyzer:
-    """Analyzes an XKB keymap to map keysyms to required modifiers"""
-    
-    def __init__(self):
-        self.context = xkb.Context()
-        self.keymap = None
-        self.layout_name = None
-        
-    def load_keymap_from_names(self, rules=None, model=None, layout=None, 
-                              variant=None, options=None):
-        """Load a keymap from RMLVO names"""
+    Lifecycle:
+        ctx = KeyboardLayoutContext(
+            apply_correction_map=lambda spec, m: keymapper.set_correction_map(m))
+        ctx.start()     # primes from the active layout, then watches for changes
+        ...
+        ctx.stop()
+
+    On start() and on every subsequent layout change, the active LayoutSpec is
+    loaded into the analyzer, compared against the reference layout, reduced to a
+    keycode->keycode correction map, and passed — together with the spec — to
+    apply_correction_map(spec, correction_map). The spec is there for observers
+    (logging, a UI indicator); a keymapper consumer ignores it and applies only
+    the map, so layout vocabulary never reaches the keymapper. The map is empty
+    for US-like layouts (no correction needed), so applying it is always safe,
+    and an unchanged layout is skipped so nothing is rebuilt needlessly.
+    """
+
+    def __init__(self, apply_correction_map, desktop_env=None,
+                    session_type=None, reference_layout='us'):
+        self._apply_correction_map = apply_correction_map
+        self._reference_layout = reference_layout
+        self._detector = KeyboardLayoutDetector(desktop_env, session_type)
+        self._analyzer = KeyboardLayoutAnalyzer()
+        self._current_spec = None
+        self._current_map = {}
+
+    @property
+    def current_layout(self):
+        """The active LayoutSpec, or None before the first resolution."""
+        return self._current_spec
+
+    @property
+    def current_correction_map(self):
+        """The correction map currently handed to the keymapper."""
+        return self._current_map
+
+    @property
+    def backend_name(self):
+        """Name of the detection backend in use (for diagnostics)."""
+        return self._detector.backend_name
+
+    def start(self) -> bool:
+        """Prime from the active layout, then watch for changes.
+
+        Returns whether a live watcher started; False means no detection backend
+        was available for this environment (nothing to coordinate).
+        """
+        # The reference (plain US) does not change, so compile it once; the
+        # analyzer keeps it across reloads of the active keymap.
+        if not self._analyzer.set_reference_from_names(layout=self._reference_layout):
+            _warn(f'could not compile reference layout {self._reference_layout!r}; '
+                    f'correction maps will be empty.')
+
+        active = self._detector.get_active_layout()
+        if active is not None:
+            self._apply_for_spec(active)
+
+        return self._detector.start(self._on_layout_change)
+
+    def stop(self):
+        """Stop watching for layout changes."""
+        self._detector.stop()
+
+    def _on_layout_change(self, spec):
+        """Detector callback; runs on the watcher thread."""
+        self._apply_for_spec(spec)
+
+    def _apply_for_spec(self, spec):
+        """Analyze one layout and hand its correction map to the keymapper."""
+        if spec == self._current_spec:
+            return                          # same layout; nothing to rebuild
+
+        if not self._analyzer.load_from_spec(spec):
+            _warn(f'could not analyze layout {format_layout(spec)}; '
+                    f'keeping the previous correction map.')
+            return
+
+        correction_map = self._analyzer.build_correction_map()
+        self._current_spec = spec
+        self._current_map = correction_map
+
+        # A consumer exception must not kill the detector's watcher thread.
         try:
-            self.keymap = self.context.keymap_new_from_names(
-                rules=rules,
-                model=model,
-                layout=layout,
-                variant=variant,
-                options=options
-            )
-            self.layout_name = self.keymap.layout_get_name(0)
-            info(f"Loaded keymap for layout: {self.layout_name}")
-        except Exception as e:
-            error(f"Failed to load keymap: {e}")
-            return False
-        return True
+            self._apply_correction_map(spec, correction_map)
+        except Exception as apply_err:
+            _warn(f'apply_correction_map raised for {format_layout(spec)}: {apply_err}')
 
-    def analyze_key(self, keycode):
-        """Analyze a single key's levels and modifiers"""
-        if not self.keymap:
-            error("No keymap loaded")
-            return None
-            
-        try:
-            key_name = self.keymap.key_get_name(keycode)
-            debug(f"\nAnalyzing key {key_name} (keycode {keycode}):")
-            
-            layout_idx = 0  # Using first layout
-            results = []
-            
-            # Get number of shift levels for this key
-            num_levels = self.keymap.num_levels_for_key(keycode, layout_idx)
-            
-            for level in range(num_levels):
-                # Get keysyms for this level
-                keysyms = self.keymap.key_get_syms_by_level(keycode, layout_idx, level)
-                symbol_names = [xkb.keysym_get_name(ks) for ks in keysyms]
-                
-                # Get modifier mask required for this level
-                mod_indices = self.keymap.key_get_mods_for_level(keycode, layout_idx, level)
-                mod_names = []
-                for idx in range(self.keymap.num_mods()):
-                    for mask in mod_indices:
-                        if mask & (1 << idx):
-                            mod_name = self.keymap.mod_get_name(idx)
-                            mod_names.append(mod_name)
-                
-                level_info = {
-                    'level': level,
-                    'symbols': symbol_names,
-                    'mod_indices': mod_indices,
-                    'mod_names': mod_names
-                }
-                results.append(level_info)
-                
-                debug(f"  Level {level}:")
-                debug(f"    Symbols: {symbol_names}")
-                debug(f"    Mod indices: {mod_indices}")
-                debug(f"    Mod names: {mod_names}")
-                
-            return results
-            
-        except xkb.XKBError as e:
-            error(f"XKB error analyzing keycode {keycode}: {e}")
-            return None
 
 def main():
-    analyzer = KeymapAnalyzer()
-    
-    # Load current layout (for testing, we'll use explicit names)
-    if not analyzer.load_keymap_from_names(layout='us', variant='mac'):
-        sys.exit(1)
-    
-    # Analyze some test keys
-    for keycode in range(9, 40):  # Test range covering main keys
-        analyzer.analyze_key(keycode)
+    """Standalone harness: announce each layout and print its correction map."""
+    def show(spec, correction_map):
+        count = len(correction_map)
+        plural = 'entry' if count == 1 else 'entries'
+        print(f'\nLayout -> {format_layout(spec)}  '
+                f'[layout={spec.layout} variant={spec.variant}]')
+        print(f'    correction map: {count} {plural}  {correction_map}')
 
-if __name__ == "__main__":
+    context = KeyboardLayoutContext(apply_correction_map=show)
+
+    print(f'Backend: {context.backend_name}')
+    print('Starting layout coordinator (Ctrl-C to stop)...')
+
+    stop_event = threading.Event()
+
+    def handle_signal(signum, frame):
+        stop_event.set()
+
+    signal.signal(signal.SIGINT, handle_signal)
+    signal.signal(signal.SIGTERM, handle_signal)
+
+    # start() primes from the active layout, so show() announces it (layout then
+    # map) before the watcher takes over; every later switch prints the same way.
+    if not context.start():
+        print('No detection backend available; nothing to coordinate.')
+        return
+
+    print('\nWatching for layout changes...')
+
+    try:
+        stop_event.wait()
+    finally:
+        context.stop()
+        print('\nStopped.')
+
+
+if __name__ == '__main__':
     main()
+
+
+# End of file #
